Bluetooth-OPP: Use long data type to contain receiving file size
If receiving file size is in GBs int type will not be able to contain
actual size and will lead to failure. So change to long instead of
int.

CRs-Fixed: 756089

+++ src/com/android/bluetooth/opp/BluetoothOppSendFileInfo.java
@@ -127,7 +127,7 @@ public class BluetoothOppSendFileInfo {
                 try {
                     if (metadataCursor.moveToFirst()) {
                         fileName = metadataCursor.getString(0);
-                        length = metadataCursor.getInt(1);
+                        length = metadataCursor.getLong(1);
                         if (D) Log.d(TAG, "fileName = " + fileName + " length = " + length);
                     }
                 } finally {
